Add files via upload
Three reference datasets for the paper "On Pseudorandomness and Deep Learning: A Case Study". Each dataset contains 5000 samples of 32-bit binary values, along with a type value. The output of a cipher or random numbers created by the Python os.urandom() function are 32-bit binary data. The name of datasets shows the related cipher (i.e., Simon and Speck) & Random numbers. One of the datasets solely contains the Simon & Speck ciphers. The value of type is either 0 or 1. The number 0 denotes the first cipher found in the dataset's name, whereas 1 indicates the other.